Move InMemoryDatabase to its own file.
No plans for reuse at this time, but it lets GameLoader.spec.ts focus on itself.
