+++ tests/database/GameLoader.spec.ts
@@ -1,120 +1,44 @@
 import {expect} from 'chai';
-import {Game, GameOptions, Score} from '../../src/Game';
+import {Game} from '../../src/Game';
 import {GameLoader} from '../../src/database/GameLoader';
 import {Player} from '../../src/Player';
 import {SerializedGame} from '../../src/SerializedGame';
+
 import {TestPlayers} from '../TestPlayers';
 import {Color} from '../../src/common/Color';
-import {GameIdLedger, IDatabase} from '../../src/database/IDatabase';
-import {GameId, PlayerId, SpectatorId} from '../../src/common/Types';
+import {GameIdLedger} from '../../src/database/IDatabase';
+import {GameId, PlayerId} from '../../src/common/Types';
 import {restoreTestDatabase, setTestDatabase} from '../utils/setup';
 import {sleep} from '../TestingUtils';
+import {InMemoryDatabase} from '../testing/InMemoryDatabase';
 
-class InMemoryDatabase implements IDatabase {
-  public data: Map<GameId, Array<SerializedGame>> = new Map();
-
+class TestDatabase extends InMemoryDatabase {
   public failure: 'getGameIds' | 'getParticipants' | undefined = undefined;
   public getGameSleep: number = 0;
 
-  initialize(): Promise<unknown> {
-    throw new Error('Method not implemented.');
+  override async getGame(gameId: GameId): Promise<SerializedGame> {
+    const game = await super.getGame(gameId);
+    await sleep(this.getGameSleep);
+    return game;
   }
 
-  async getGame(gameId: GameId): Promise<SerializedGame> {
-    const row = this.data.get(gameId);
-    if (row === undefined || row.length === 0) {
-      throw new Error('not found');
-    } else {
-      const game = row[row.length -1];
-      await sleep(this.getGameSleep);
-      return game;
-    }
-  }
-  getGameId(_id: PlayerId | SpectatorId): Promise<GameId> {
-    throw new Error('Method not implemented.');
-  }
-  getSaveIds(gameId: GameId): Promise<number[]> {
-    const row = this.data.get(gameId);
-    if (row === undefined || row.length === 0) {
-      return Promise.reject(new Error('not found'));
-    } else {
-      return Promise.resolve([...Array(row?.length).keys()]);
-    }
-  }
-  getGameVersion(gameId: GameId, saveId: number): Promise<SerializedGame> {
-    const row = this.data.get(gameId);
-    if (row === undefined || row.length === 0) {
-      return Promise.reject(new Error('not found'));
-    }
-    const serializedGame = row[saveId];
-    if (serializedGame === undefined) {
-      return Promise.reject(new Error('not found'));
-    }
-    return Promise.resolve(serializedGame);
-  }
-  getGameIds(): Promise<GameId[]> {
+  override getGameIds(): Promise<GameId[]> {
     if (this.failure === 'getGameIds') return Promise.reject(new Error('error'));
-    return Promise.resolve(Array.from(this.data.keys()));
+    return super.getGameIds();
   }
-  getPlayerCount(_game_id: string): Promise<number> {
-    throw new Error('Method not implemented.');
-  }
-  saveGame(game: Game): Promise<void> {
-    const gameId = game.id;
-    const row = this.data.get(gameId) || [];
-    this.data.set(gameId, row);
-    while (row.length < game.lastSaveId) {
-      row.push();
-    }
-    row[game.lastSaveId] = game.serialize();
-    game.lastSaveId++;
-    return Promise.resolve();
-  }
-  saveGameResults(_game_id: string, _players: number, _generations: number, _gameOptions: GameOptions, _scores: Score[]): void {
-    throw new Error('Method not implemented.');
-  }
-  restoreGame(_game_id: string, _save_id: number): Promise<SerializedGame> {
-    throw new Error('Method not implemented.');
-  }
-  loadCloneableGame(_game_id: string): Promise<SerializedGame> {
-    throw new Error('Method not implemented.');
-  }
-  deleteGameNbrSaves(_game_id: string, _rollbackCount: number): void {
-    throw new Error('Method not implemented.');
-  }
-  cleanGame(_game_id: string): Promise<void> {
-    throw new Error('Method not implemented.');
-  }
-  purgeUnfinishedGames(): Promise<void> {
-    throw new Error('Method not implemented.');
-  }
-  stats(): Promise<{[ key: string ]: string | number;}> {
-    throw new Error('Method not implemented.');
-  }
-  storeParticipants() {
-    return Promise.resolve();
-  }
-  async getParticipants(): Promise<Array<GameIdLedger>> {
+  override getParticipants(): Promise<Array<GameIdLedger>> {
     if (this.failure === 'getParticipants') return Promise.reject(new Error('error'));
-
-    const entries: Array<GameIdLedger> = [];
-    this.data.forEach((games, gameId) => {
-      const firstSave = games[0];
-      const participantIds: Array<PlayerId | SpectatorId> = firstSave.players.map((p) => p.id);
-      if (firstSave.spectatorId) participantIds.push(firstSave.spectatorId);
-      entries.push({gameId, participantIds});
-    });
-    return entries;
+    return super.getParticipants();
   }
 }
 
 describe('GameLoader', function() {
   let instance: GameLoader;
-  let database: InMemoryDatabase;
+  let database: TestDatabase;
   let game: Game;
 
   beforeEach(function() {
-    database = new InMemoryDatabase();
+    database = new TestDatabase();
     setTestDatabase(database);
     const player = TestPlayers.BLUE.newPlayer();
     const player2 = TestPlayers.RED.newPlayer();
